requirements.txt
requests==2.3.0
PyYAML==3.11
